Reorganize source code into logical folder structure
**Directory Organization:**
- config/: Configuration files and data types
  - config.h (main config aggregator)
  - system_config.h, sensor_config.h, alert_config.h
  - data_types.h (SystemState, SensorReadings structures)

- sensors/: Sensor hardware management
  - sensor_manager.h/.cpp (I2C sensor coordination)

- data_processing/: Data analysis and algorithms
  - data_processor.h/.cpp (main coordinator)
  - anomaly_detector.h/.cpp (anomaly detection algorithms)
  - statistical_analyzer.h/.cpp (statistical analysis and trends)

- communication/: External communication
  - notecard_manager.h/.cpp (cellular IoT communication)
  - telemetry_formatter.h/.cpp (JSON data formatting)

- alerts/: Alert management system
  - alert_handler.h/.cpp (alert processing and routing)

- utils/: Utility functions and cross-cutting concerns
  - error_handling.h/.cpp (error management system)

**Benefits:**
- Clear separation of concerns by functionality
- Easier navigation and maintenance
- Logical grouping of related components
- Improved code discoverability
- Better encapsulation of module responsibilities
- Maintains all existing functionality

**Updated Include Paths:**
- Updated all cross-module references with relative paths
- Maintained build compatibility
- All tests pass successfully
