+++ conveyor_monitor/src/communication/notecard_manager.cpp
