+++ conveyor_monitor/src/config/sensor_config.h
