+++ conveyor_monitor/src/data_processing/anomaly_detector.cpp
