+++ conveyor_monitor/src/data_processing/anomaly_detector.h
@@ -2,8 +2,8 @@
 #define ANOMALY_DETECTOR_H
 
 #include <Arduino.h>
-#include "data_types.h"
-#include "sensor_config.h"
+#include "../config/data_types.h"
+#include "../config/sensor_config.h"
 
 /**
  * @brief Specialized class for detecting system anomalies
