+++ conveyor_monitor/src/data_processing/data_processor.cpp
