+++ conveyor_monitor/src/sensors/sensor_manager.h
@@ -2,7 +2,7 @@
 #define SENSOR_MANAGER_H
 
 #include <Arduino.h>
-#include "config.h"
+#include "../config/config.h"
 
 class SensorManager {
 private:
