+++ conveyor_monitor/src/utils/error_handling.cpp
